+++ Utility.h
@@ -15,8 +15,8 @@ class Utility {
     static void Init();
     static int Random(int min, int max);
     static double Random();
-    static float pow(float base, int exp);
-    static std::vector<double> Random(unsigned int count); // return a float number between 0 and 1
+    static double pow(double base, int exp);
+    static std::vector<double> Random(unsigned int count); // return a double number between 0 and 1
 };
